+++ src/Octonica.ClickHouseClient/ClickHouseDbType.cs
@@ -1,5 +1,5 @@
 ﻿#region License Apache 2.0
-/* Copyright 2020-2021 Octonica
+/* Copyright 2020-2021, 2024 Octonica
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -206,6 +206,7 @@ namespace Octonica.ClickHouseClient
 
         /// <summary>
         /// A type representing a date and time value with defined sub-second precision.
+        /// Supported range of values: [1900-01-01 00:00:00, 2299-12-31 23:59:59.99999999]
         /// </summary>
         DateTime64 = ClickHouseSpecificTypeDelimiterCode + 7,
 
@@ -236,8 +237,14 @@ namespace Octonica.ClickHouseClient
 
         /// <summary>
         /// A type representing a date value without a time. Supports the date range same with <see cref="DateTime64"/>.
-        /// Stored in four bytes as the number of days since 1925-01-01. Allows storing values till 2283-11-11.
+        /// Stored in four bytes as the number of days since 1970-01-01.
         /// </summary>
         Date32 = ClickHouseSpecificTypeDelimiterCode + 13,
+
+        /// <summary>
+        /// This type represents a union of other data types. Type Variant(T1, T2, ..., TN) means that each row of this type
+        /// has a value of either type T1 or T2 or ... or TN or none of them (NULL value).
+        /// </summary>
+        Variant = ClickHouseSpecificTypeDelimiterCode + 14,
     }
 }