+++ src/Octonica.ClickHouseClient/Types/ClickHouseTypeInfoProvider.cs
@@ -1,5 +1,5 @@
 ﻿#region License Apache 2.0
-/* Copyright 2019-2022 Octonica
+/* Copyright 2019-2022, 2024 Octonica
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -616,6 +616,7 @@ namespace Octonica.ClickHouseClient.Types
                 new Enum16TypeInfo(),
 
                 new MapTypeInfo(),
+                new VariantTypeInfo(),
             };
         }
     }